구름 준비해놨고, fmod 매니저 제작

+++ Hungry_Dragon/Client/Code/Cloud_Locater.cpp
@@ -83,8 +83,8 @@ void CCloud_Locater::Render_Object(void)
 	{
 		for (int j = 0; j < 3; ++j)
 		{
-			m_pCloud[_iIdx]->Set_Trans(_vec3((int)(m_uLength*(m_uVerCntX-1))*(-1+j), 
-										-5000, (int)(m_uLength*(m_uVerCntZ - 1))*(iPlayerZ-1+i)));
+			m_pCloud[_iIdx]->Set_Trans(_vec3((float)((int)(m_uLength*(m_uVerCntX-1))*(-1+j)),
+										-5000, (float)((int)(m_uLength*(m_uVerCntZ - 1))*(iPlayerZ-1+i))));
 			m_pCloud[_iIdx]->Render_Object();
 			_iIdx = (_iIdx+1)%9;
 		}
@@ -95,8 +95,8 @@ void CCloud_Locater::Render_Object(void)
 	{
 		for (int j = 0; j < 3; ++j)
 		{
-			m_pCloud[_iIdx]->Set_Trans(_vec3((int)(m_uLength*(m_uVerCntX - 1))*( - 1 + j),
-				5000, (int)(m_uLength*(m_uVerCntZ - 1))*(iPlayerZ - 1 + i)));
+			m_pCloud[_iIdx]->Set_Trans(_vec3((float)((int)(m_uLength*(m_uVerCntX - 1))*( - 1 + j)),
+				5000, (float)((int)(m_uLength*(m_uVerCntZ - 1))*(iPlayerZ - 1 + i))));
 			m_pCloud[_iIdx]->Render_Object();
 			_iIdx = (_iIdx + 1) % 9;
 		}
